requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
tenacity>=8.2.3
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.30
# Langfuse is only required if your original code imports it
langfuse>=2.0.0

